+++ test/main.js
@@ -9,6 +9,7 @@ import "/test/api/labels/labels.test";
 import "/test/api/notifications/notifications.test";
 import "/test/api/canvas/canvas.test";
 import "/test/api/healthReports/healthReports.test";
+import "/test/api/dashboards/dashboards.test";
 
 describe("latelier", function() {
   it("package.json has correct name", async function() {